requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
google-api-python-client==2.34.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6